moved review and add. info to own components
